// src/types/ErrorCodes.ts
export enum ErrorCodes {
    // 1000 - 1999: General errors

    INVALID_REQUEST = 1000,
    UNAUTHORIZED = 1001,
    FORBIDDEN = 1002,
    NOT_FOUND = 1003,
    METHOD_NOT_ALLOWED = 1004,
    INTERNAL_SERVER_ERROR = 1005,

    // 2000 - 2999: Authentication errors

    INVALID_TOKEN = 2000,
    EXPIRED_TOKEN = 2001,
    INVALID_CREDENTIALS = 2002,
    USER_NOT_FOUND = 2003,
    USER_ALREADY_EXISTS = 2004,
    INVALID_VERIFICATION_CODE = 2007,
    INVALID_REFRESH_TOKEN = 2008,
    INVALID_2FA_CODE = 2009,
    USER_NOT_VERIFIED = 2010,

    // 3000 - 3999: Database errors

    DATABASE_ERROR = 3000,
    DUPLICATE_ENTRY = 3001,
    INVALID_QUERY = 3002,
    CONNECTION_ERROR = 3003,
    TIMEOUT_ERROR = 3004,
    NOT_FOUND_IN_DATABASE = 3005,
    INVALID_ID = 3006,

    // 4000 - 4999: Validation errors

    MISSING_FIELDS = 4000,
}